Fix bugs and update user interface
A bug that was caused by multiple operation in sequence (2++2...etc) was
fixed.

A bug that was caused by pressing the enter key and inserting a new line
 of blank text was fixed.

A bug that was caused by multiple decimal points in a single number was
fixed.

A bug that was caused by adjacent parenthesis ((2)(2)...etc) was fixed.

A bug that was caused by blank parenthesis was fixed.